API-Character-Sprite-Generator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
pyppeteer==1.0.2
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pillow==10.4.0
numpy==1.26.4